+++ fractal_server/app/routes/api/v2/task_collection.py
@@ -34,7 +34,7 @@ from fractal_server.app.schemas.v2 import (
     TaskGroupActivityActionV2,
 )
 from fractal_server.app.schemas.v2 import TaskGroupV2OriginEnum
-from fractal_server.tasks.v2.collection_local import (
+from fractal_server.tasks.v2.local.collect import (
     collect_package_local,
 )
 from fractal_server.tasks.v2.utils_package_names import _parse_wheel_filename
@@ -247,7 +247,7 @@ async def collect_tasks_pip(
     if settings.FRACTAL_RUNNER_BACKEND == "slurm_ssh":
         # SSH task collection
 
-        from fractal_server.tasks.v2.collection_ssh import (
+        from fractal_server.tasks.v2.ssh.collect import (
             collect_package_ssh,
         )
 
@@ -262,8 +262,8 @@ async def collect_tasks_pip(
 
         background_tasks.add_task(
             collect_package_ssh,
+            task_group_id=task_group.id,
             task_group_activity_id=task_group_activity.id,
-            task_group=task_group,
             fractal_ssh=fractal_ssh,
             tasks_base_dir=user_settings.ssh_tasks_dir,
         )
@@ -272,8 +272,8 @@ async def collect_tasks_pip(
         # Local task collection
         background_tasks.add_task(
             collect_package_local,
+            task_group_id=task_group.id,
             task_group_activity_id=task_group_activity.id,
-            task_group=task_group,
         )
     logger.debug(
         "Task-collection endpoint: start background collection "