+++ tests/v2/06_tasks/test_background_operations_v2.py
@@ -9,11 +9,12 @@ from fractal_server.app.models.v2 import TaskGroupV2
 from fractal_server.app.schemas.v2 import (
     TaskGroupActivityStatusV2,
 )
-from fractal_server.tasks.v2.collection_local import collect_package_local
-from fractal_server.tasks.v2.database_operations import _get_task_type
+from fractal_server.tasks.v2.local.collect import collect_package_local
+from fractal_server.tasks.v2.ssh.collect import collect_package_ssh
 from fractal_server.tasks.v2.utils_background import (
     check_task_files_exist,
 )
+from fractal_server.tasks.v2.utils_database import _get_task_type
 
 
 class _MockTaskCreateV2(BaseModel):
@@ -93,7 +94,7 @@ async def test_collect_pip_existing_file(tmp_path, db, first_user):
     path.mkdir()
     # Run background task
     collect_package_local(
-        task_group=task_group,
+        task_group_id=task_group.id,
         task_group_activity_id=task_group_activity.id,
     )
     # Verify that collection failed
@@ -114,13 +115,13 @@ async def test_collect_pip_local_fail_rmtree(
     monkeypatch,
 ):
 
-    import fractal_server.tasks.v2.collection_local
+    import fractal_server.tasks.v2.local.collect
 
     def patched_function(*args, **kwargs):
         raise RuntimeError("Broken rm")
 
     monkeypatch.setattr(
-        fractal_server.tasks.v2.collection_local.shutil,
+        fractal_server.tasks.v2.local.collect.shutil,
         "rmtree",
         patched_function,
     )
@@ -164,7 +165,7 @@ async def test_collect_pip_local_fail_rmtree(
     # Run background task
     try:
         collect_package_local(
-            task_group=task_group,
+            task_group_id=task_group.id,
             task_group_activity_id=task_group_activity.id,
         )
     except RuntimeError as e:
@@ -180,3 +181,26 @@ async def test_collect_pip_local_fail_rmtree(
     assert task_group_activity_v2.status == "failed"
     assert "Broken rm" in task_group_activity_v2.log
     assert path.exists()
+
+
+def test_unit_missing_objects(db, caplog):
+    """
+    Test a branch which is in principle unreachable.
+    """
+    caplog.clear()
+    collect_package_local(
+        task_group_activity_id=9999,
+        task_group_id=9999,
+    )
+    assert "Cannot find database rows" in caplog.text
+
+    caplog.clear()
+    assert caplog.text == ""
+
+    collect_package_ssh(
+        task_group_activity_id=9999,
+        task_group_id=9999,
+        fractal_ssh=None,
+        tasks_base_dir="/invalid",
+    )
+    assert "Cannot find database rows" in caplog.text